rag_common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-common"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "numpy>=1.24",
    "faiss-cpu>=1.7",
    "rank-bm25>=0.2",
    "nltk>=3.8",
    "sentence-transformers>=2.2",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[tool.hatch.build.targets.wheel]
packages = ["rag_common"]
